odooLinux/requirements.txt
click>=8.0.0
debugpy>=1.6.0
psutil>=5.9.0
weasyprint>=60.0
pydyf>=0.8.0
speedtest-cli>=2.1.3
requests>=2.31.0
tqdm>=4.65.0
pandas
matplotlib
pdfkit 